Move logger library to commons
