+++ components/application-connectivity-validator/cmd/applicationconnectivityvalidator/applicationconnectivityvalidator.go
@@ -8,11 +8,11 @@ import (
 	"time"
 
 	"github.com/kyma-project/kyma/components/application-connectivity-validator/internal/controller"
-	"github.com/kyma-project/kyma/components/application-connectivity-validator/pkg/tracing"
+	"github.com/kyma-project/kyma/common/logger/tracing"
 
 	"github.com/kyma-project/kyma/components/application-connectivity-validator/internal/externalapi"
 	"github.com/kyma-project/kyma/components/application-connectivity-validator/internal/validationproxy"
-	logger "github.com/kyma-project/kyma/components/application-connectivity-validator/pkg/logger"
+	logger "github.com/kyma-project/kyma/common/logger/logger"
 	"github.com/patrickmn/go-cache"
 )
