+++ components/application-connectivity-validator/go.mod
@@ -13,11 +13,12 @@ require (
 	github.com/hashicorp/golang-lru v0.5.4 // indirect
 	github.com/imdario/mergo v0.3.9 // indirect
 	github.com/kr/text v0.2.0 // indirect
+	github.com/kyma-project/kyma/common v0.0.0-00010101000000-000000000000
 	github.com/kyma-project/kyma/components/application-operator v0.0.0-20200902071617-83c683010f30
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/patrickmn/go-cache v2.1.0+incompatible
 	github.com/stretchr/testify v1.6.1
-	go.uber.org/zap v1.13.0
+	go.uber.org/zap v1.16.0
 	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d // indirect
 	golang.org/x/time v0.0.0-20200416051211-89c76fbcd5d1 // indirect
 	golang.org/x/tools v0.0.0-20200924205911-8a9a89368bd3 // indirect
@@ -28,12 +29,13 @@ require (
 	k8s.io/api v0.18.8 // indirect
 	k8s.io/apimachinery v0.18.8
 	k8s.io/client-go v0.18.8
-	k8s.io/klog/v2 v2.2.0
+	k8s.io/klog/v2 v2.4.0
 )
 
 replace (
 	github.com/containerd/containerd => github.com/containerd/containerd v1.4.1
 	github.com/docker/docker => github.com/docker/docker v17.12.0-ce-rc1.0.20200916142827-bd33bbf0497b+incompatible
+	github.com/kyma-project/kyma/common => github.com/franpog859/kyma/common v0.0.0-20210121091704-7ae5af7db5e8
 	github.com/opencontainers/runc => github.com/opencontainers/runc v1.0.0-rc92
 	golang.org/x/crypto => golang.org/x/crypto v0.0.0-20200604202706-70a84ac30bf9
 	golang.org/x/text => golang.org/x/text v0.3.3
