+++ components/application-connectivity-validator/internal/validationproxy/handler_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/kyma-project/kyma/components/application-connectivity-validator/pkg/logger"
+	"github.com/kyma-project/kyma/common/logger/logger"
 
 	"github.com/gorilla/mux"
 	"github.com/patrickmn/go-cache"
